@@ -1,9 +1,8 @@
-﻿namespace MaksimShimshon.RestCountries.DataModels
+﻿namespace MaksimShimshon.RestCountries.DataModels;
+
+public record class CountryNameDM
 {
-    public record class CountryNameDM
-    {
-        public string Common { get; init; } = null!;
-        public string Official { get; init; } = null!;
-        public Dictionary<string, CountryNameTranslatedDM> NativeName { get; init; } = null!;
-    }
+    public string Common { get; init; } = null!;
+    public string Official { get; init; } = null!;
+    public Dictionary<string, CountryNameTranslatedDM> NativeName { get; init; } = null!;
 }
