@@ -1,9 +1,8 @@
-﻿namespace MaksimShimshon.RestCountries.Entities
+﻿namespace MaksimShimshon.RestCountries.Entities;
+
+public record class CountryDemonyms
 {
-    public record class CountryDemonyms
-    {
-        public string LanguageCode { get; init; } = null!;
-        public string Female { get; init; } = null!;
-        public string Male { get; init; } = null!;
-    }
+    public string LanguageCode { get; init; } = null!;
+    public string Female { get; init; } = null!;
+    public string Male { get; init; } = null!;
 }
