@@ -1,8 +1,7 @@
-﻿namespace MaksimShimshon.RestCountries.Entities
+﻿namespace MaksimShimshon.RestCountries.Entities;
+
+public record class CountryPostalFormat
 {
-    public record class CountryPostalFormat
-    {
-        public string Format { get; init; } = null!;
-        public string Regex { get; init; } = null!;
-    }
+    public string Format { get; init; } = null!;
+    public string Regex { get; init; } = null!;
 }
