@@ -1,25 +1,24 @@
-﻿namespace MaksimShimshon.RestCountries.Tests.Engine
+﻿namespace MaksimShimshon.RestCountries.Tests.Engine;
+
+public class EngineCurrencyCCA2Tests : EngineTests
 {
-    public class EngineCurrencyCCA2Tests : EngineTests
+
+    [Theory]
+    [InlineData("AU", "AUD", "countriesV3.1")]
+    [InlineData("ca", "CAD", "countriesV3")]
+    public void IsValidCCA2Currency_Success(string countryCode, string currencyCode, string version)
     {
-
-        [Theory]
-        [InlineData("AU", "AUD", "countriesV3.1")]
-        [InlineData("ca", "CAD", "countriesV3")]
-        public void IsValidCCA2Currency_Success(string countryCode, string currencyCode, string version)
-        {
-            Assert.True(GetOrCache(version).IsValidCCA2Currency(countryCode, currencyCode));
-        }
-
-
-        [Theory]
-        [InlineData("AU", "CAD", "countriesV3.1")]
-        [InlineData("ca", "USD", "countriesV3")]
-        public void IsValidCCA2Currency_Failures(string countryCode, string currencyCode, string version)
-        {
-            Assert.False(GetOrCache(version).IsValidCCA2Currency(countryCode, currencyCode));
-        }
-
-
+        Assert.True(GetOrCache(version).IsValidCCA2Currency(countryCode, currencyCode));
     }
+
+
+    [Theory]
+    [InlineData("AU", "CAD", "countriesV3.1")]
+    [InlineData("ca", "USD", "countriesV3")]
+    public void IsValidCCA2Currency_Failures(string countryCode, string currencyCode, string version)
+    {
+        Assert.False(GetOrCache(version).IsValidCCA2Currency(countryCode, currencyCode));
+    }
+
+
 }
